Change the read property when Node read is changed
Added a function that is binded to the Book.prototype
and a function with event listener binded to the
created book. When the toggleRead is run, also change
the internal state of read status. This could be
done bettter, for example better divide the internal
functions and the functions for displaying books